refactored code to split out mapping and fh and added unit tests
