+++ src/test/controllers/order.confirmation.controller.integration.test.ts
@@ -1,22 +1,22 @@
 import chai from "chai";
 import sinon from "sinon";
 import ioredis from "ioredis";
-import { Order } from "ch-sdk-node/dist/services/order/order";
-import { Basket } from "ch-sdk-node/dist/services/order/basket";
+import { Basket } from "api-sdk-node/dist/services/order/basket";
 import cheerio from "cheerio";
 
 import * as apiClient from "../../client/api.client";
 import { SIGNED_IN_COOKIE, signedInSession } from "../__mocks__/redis.mocks";
+import {
+    ORDER_ID, mockCertificateOrderResponse, mockCertifiedCopyOrderResponse,
+    mockMissingImageDeliveryOrderResponse
+} from "../__mocks__/order.mocks";
 
 const sandbox = sinon.createSandbox();
 let testApp = null;
 let getOrderStub;
 let getBasketStub;
 
-const ORDER_ID = "ORD-123456-123456";
 const ORDER_ID_ARIA_LABEL = "ORD hyphen 123456 hyphen 123456";
-const CERTIFICATE_ID = "CRT-123456-123456";
-const CERTIFIED_COPY_ID = "CCD-123456-123456";
 
 const ITEM_KINDS = [{
     kind: "item#certificate",
@@ -28,159 +28,6 @@ const ITEM_KINDS = [{
     name: "certified-copy",
     url: "/orderable/certified-copies/CCD-123456-123456"
 }];
-
-const mockCertificateOrderResponse: Order = {
-    orderedAt: "2019-12-16T09:16:17.791Z",
-    orderedBy: {
-        id: "123456",
-        email: "email@examlpe.come"
-    },
-    links: {
-        self: `/orders/${ORDER_ID}`
-    },
-    paymentReference: "1234567",
-    etag: "abcdefghijk123456789",
-    deliveryDetails: {
-        addressLine1: "address line 1",
-        addressLine2: "address line 2",
-        country: "country",
-        forename: "forename",
-        locality: "locality",
-        postalCode: "postal code",
-        region: "region",
-        surname: "surname",
-        poBox: "po box"
-    },
-    items: [{
-        id: CERTIFICATE_ID,
-        companyName: "Company Name",
-        companyNumber: "00000000",
-        description: "certificate for company 00000000",
-        descriptionIdentifier: "certificate",
-        descriptionValues: {
-            certificate: "certificate for company 00000000",
-            companyNumber: "00000000"
-        },
-        itemCosts: [{
-            discountApplied: "0",
-            itemCost: "15",
-            calculatedCost: "15",
-            productType: "certificate"
-        }],
-        itemOptions: {
-            certificateType: "incorporation-with-all-name-changes",
-            deliveryMethod: "postal",
-            deliveryTimescale: "standard",
-            directorDetails: {},
-            forename: "forename",
-            includeGoodStandingInformation: true,
-            registeredOfficeAddressDetails: {},
-            secretaryDetails: {},
-            surname: "surname"
-        },
-        etag: "abcdefg123456",
-        kind: "item#certificate",
-        links: {
-            self: "/orderable/certificates/" + CERTIFICATE_ID
-        },
-        postalDelivery: true,
-        quantity: 1,
-        itemUri: "/orderable/certificates/" + CERTIFICATE_ID,
-        status: "unknown",
-        postageCost: "0",
-        totalItemCost: "15",
-        customerReference: "mycert",
-        satisfiedAt: "2020-05-15T08:41:05.798Z"
-    }
-    ],
-    kind: "order",
-    totalOrderCost: "15",
-    reference: ORDER_ID
-};
-
-const mockCertifiedCopyOrderResponse: Order = {
-    orderedAt: "2019-12-16T09:16:17.791Z",
-    orderedBy: {
-        id: "123456",
-        email: "email@examlpe.come"
-    },
-    links: {
-        self: `/orders/${ORDER_ID}`
-    },
-    paymentReference: "1234567",
-    etag: "abcdefghijk123456789",
-    deliveryDetails: {
-        addressLine1: "address line 1",
-        addressLine2: "address line 2",
-        country: "country",
-        forename: "forename",
-        locality: "locality",
-        postalCode: "postal code",
-        region: "region",
-        surname: "surname",
-        poBox: "po box"
-    },
-    items: [{
-        id: CERTIFICATE_ID,
-        companyName: "Company Name",
-        companyNumber: "00000000",
-        description: "certified-copy for company 00000000",
-        descriptionIdentifier: "certified-copy",
-        descriptionValues: {
-            certificate: "certified-copy for company 00000000",
-            companyNumber: "00000000"
-        },
-        itemCosts: [{
-            discountApplied: "0",
-            itemCost: "30",
-            calculatedCost: "30",
-            productType: "certified-copy"
-        }],
-        itemOptions: {
-            deliveryMethod: "postal",
-            deliveryTimescale: "standard",
-            filingHistoryDocuments: [{
-                filingHistoryDate: "2010-02-12",
-                filingHistoryDescription: "change-person-director-company-with-change-date",
-                filingHistoryDescriptionValues: {
-                    change_date: "2010-02-12",
-                    officer_name: "Thomas David Wheare"
-                },
-                filingHistoryId: "MzAwOTM2MDg5OWFkaXF6a2N4",
-                filingHistoryType: "CH01",
-                filingHistoryCost: "15"
-            },
-            {
-                filingHistoryDate: "2009-03-12",
-                filingHistoryDescription: "accounts-with-accounts-type-group",
-                filingHistoryDescriptionValues: {
-                    made_up_date: "2008-08-31"
-                },
-                filingHistoryId: "MjAzNTYyNTE5M2FkaXF6a2N4",
-                filingHistoryType: "AA",
-                filingHistoryCost: "15"
-            }
-            ]
-        },
-        etag: "abcdefg123456",
-        kind: "item#certified-copy",
-        links: {
-            self: "/orderable/certified-copies/" + CERTIFIED_COPY_ID
-        },
-        postalDelivery: true,
-        quantity: 1,
-        itemUri: "/orderable/certified-copies/" + CERTIFIED_COPY_ID,
-        status: "unknown",
-        postageCost: "0",
-        totalItemCost: "30",
-        customerReference: "mycert",
-        satisfiedAt: "2020-05-15T08:41:05.798Z"
-    }
-    ],
-    kind: "order",
-    totalOrderCost: "30",
-    reference: ORDER_ID
-};
 
 describe("order.confirmation.controller.integration", () => {
     beforeEach(done => {
@@ -251,6 +98,31 @@ describe("order.confirmation.controller.integration", () => {
         chai.expect($("#paymentAmountValue").text()).to.equal("£30");
         chai.expect($("#paymentReferenceValue").text()).to.equal(mockCertifiedCopyOrderResponse.paymentReference);
         chai.expect($("#paymentTimeValue").text()).to.equal("16 December 2019 - 09:16:17");
+        chai.expect(resp.text).to.not.contain("certificateTypeValue");
+        chai.expect(resp.text).to.not.contain("includedOnCertificateValue");
+        chai.expect(getOrderStub).to.have.been.called;
+    });
+
+    it("renders get order page on successful get order call for a missing image delivery order", async () => {
+        getOrderStub = sandbox.stub(apiClient, "getOrder").returns(Promise.resolve(mockMissingImageDeliveryOrderResponse));
+
+        const resp = await chai.request(testApp)
+            .get(`/orders/${ORDER_ID}/confirmation?ref=orderable_item_${ORDER_ID}&state=1234&status=paid&itemType=missing-image-delivery`)
+            .set("Cookie", [`__SID=${SIGNED_IN_COOKIE}`]);
+
+        const $ = cheerio.load(resp.text);
+
+        chai.expect(resp.status).to.equal(200);
+        chai.expect($("#orderReference").text()).to.equal(mockMissingImageDeliveryOrderResponse.reference);
+        chai.expect($("#orderReference").attr("aria-label")).to.equal(ORDER_ID_ARIA_LABEL);
+        chai.expect($("#companyNameValue").text()).to.equal(mockMissingImageDeliveryOrderResponse.items[0].companyName);
+        chai.expect($("#companyNumberValue").text()).to.equal(mockMissingImageDeliveryOrderResponse.items[0].companyNumber);
+        chai.expect($("#filingHistoryDateValue").text().trim()).to.equal("26 May 2015");
+        chai.expect($("#filingHistoryTypeValue").text().trim()).to.equal("AP01");
+        chai.expect($("#filingHistoryDescriptionValue").text().trim()).to.equal("Appointment of Mr Richard John Harris as a director");
+        chai.expect($("#paymentAmountValue").text()).to.equal("£3");
+        chai.expect($("#paymentReferenceValue").text()).to.equal(mockMissingImageDeliveryOrderResponse.paymentReference);
+        chai.expect($("#paymentTimeValue").text()).to.equal("07 October 2020 - 11:09:46");
         chai.expect(resp.text).to.not.contain("certificateTypeValue");
         chai.expect(resp.text).to.not.contain("includedOnCertificateValue");
         chai.expect(getOrderStub).to.have.been.called;
