+++ src/views/order-complete.html
@@ -7,7 +7,7 @@
 {% block beforeContent %}
   {{ govukBackLink({
     text: "Return to company overview",
-    href: "/company/" + itemDetails.companyNumber
+    href: "/company/" + companyNumber
   }) }}
 {% endblock %}
